+++ lib/quantum/matcher.ex
@@ -8,16 +8,16 @@ defmodule Quantum.Matcher do
     true
   end
 
-  def match([], _, _, _) do
+  def match(e, v, min, max) do
+    do_match(e |> String.split(","), v, min, max)
+  end
+
+  defp do_match([], _, _, _) do
     false
   end
 
-  def match([e|t], v, min, max) do
-    Enum.any?(parse(e, min, max), &(&1 == v)) or match(t, v, min, max)
-  end
-
-  def match(e, v, min, max) do
-    match(e |> String.split(","), v, min, max)
+  defp do_match([e|t], v, min, max) do
+    Enum.any?(parse(e, min, max), &(&1 == v)) or do_match(t, v, min, max)
   end
 
 end