+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.2.2] - 2015-06-15
 ### Added
 - Support for `@annually` and `@midnight`
 
@@ -84,7 +86,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Initial commit
 
 
-[unreleased]: https://github.com/c-rack/quantum-elixir/compare/v1.2.1...HEAD
+[unreleased]: https://github.com/c-rack/quantum-elixir/compare/v1.2.2...HEAD
+[1.2.2]: https://github.com/c-rack/quantum-elixir/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/c-rack/quantum-elixir/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/c-rack/quantum-elixir/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/c-rack/quantum-elixir/compare/v1.0.4...v1.1.0